Added exception handling for null entities with testing.

+++ src/main/java/com/example/demo/controllers/CustomerController.java
@@ -1,10 +1,12 @@
 package com.example.demo.controllers;
 
 import com.example.demo.entities.Customer;
+import com.example.demo.exception.NotFoundException;
 import com.example.demo.services.CustomerService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.*;
 
@@ -29,12 +31,22 @@ public class CustomerController {
     @ResponseBody
     @GetMapping("/{id}")
     public Customer findCustomerById(@PathVariable("id") Long id) {
-        return customerService.findCustomerById(id);
+        Customer customer = customerService.findCustomerById(id);
+        if (customer == null) {
+            throw new NotFoundException();
+        } else {
+            return customer;
+        }
     }
 
     @DeleteMapping("/{id}")
     public void deleteCustomer(@PathVariable("id") Long id) {
-        customerService.deleteCustomer(id);
+        Customer customer = findCustomerById(id);
+        if(customer == null) {
+            throw new NotFoundException();
+        } else {
+            customerService.deleteCustomer(id);
+        }
     }
 
 }
