+++ src/main/java/com/example/demo/controllers/LocationController.java
@@ -1,6 +1,7 @@
 package com.example.demo.controllers;
 
 import com.example.demo.entities.Location;
+import com.example.demo.exception.NotFoundException;
 import com.example.demo.services.LocationService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
@@ -29,11 +30,21 @@ public class LocationController {
     @ResponseBody
     @GetMapping(path = "/{id}")
     public Location findLocationById(@PathVariable("id") Long id){
-        return locationService.findLocationById(id);
+        Location location = locationService.findLocationById(id);
+        if (location == null) {
+            throw new NotFoundException();
+        } else {
+            return locationService.findLocationById(id);
+        }
     }
 
     @DeleteMapping("/{id}")
     public void deleteLocation(@PathVariable("id") long id){
-        locationService.deleteLocation(id);
+        Location location = locationService.findLocationById(id);
+        if (location == null) {
+            throw new NotFoundException();
+        } else {
+            locationService.deleteLocation(id);
+        }
     }
 }
